% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigcurve-package.R
\docType{package}
\name{bigcurve-package}
\alias{bigcurve}
\alias{bigcurve-package}
\title{bigcurve: Adaptive Densifying of Lines}
\description{
Densify lines adaptively, as required by the curvature a map projection imposes on them. Vertices are inserted along great circle arcs only where the projected curve deviates from a straight chord by more than a tolerance, following the adaptive resampling scheme of 'd3-geo' by Mike Bostock.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/bigcurve}
  \item Report bugs at \url{https://github.com/hypertidy/bigcurve/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

Authors:
\itemize{
  \item Michael D. Sumner \email{mdsumner@gmail.com}
}

Other contributors:
\itemize{
  \item Simon Wotherspoon (wrote SGAT::trackMdpts, the ancestor of the spherical midpoint used here) [contributor]
}

}
\keyword{internal}
