% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisect.R
\name{curv_len}
\alias{curv_len}
\title{Curvature index for a given segment}
\usage{
curv_len(x, proj)
}
\arguments{
\item{x}{linestring, single segment}

\item{proj}{projection string (proj, wkt, auth, whatever)}
}
\value{
distance in metres
}
\description{
Calculate the perpendicular distance from the mid point of the projected
version of a segment to its arc mid-point.
}
\details{
Yes this is the D3 Flawed Example logic.
}
\examples{
s1 <- segment(c(0, 1), c(0, 1))
curv_len(s1, laea(147, -42))
curv_len(s1, laea(100, -30))
curv_len(s1, laea(70, -20))
curv_len(s1, laea(30, -10))
curv_len(s1, laea(0, 0))
}
\references{
https://bost.ocks.org/mike/example/
}
