% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{laea}
\alias{laea}
\title{Projection string}
\usage{
laea(lon_0 = 0, lat_0 = 0, x_0 = 0, y_0 = 0)
}
\arguments{
\item{lon_0}{centre longitude}

\item{lat_0}{centre latitude}

\item{x_0}{false easting (default 0 is fine)}

\item{y_0}{false northing (default 0 is fine)}
}
\value{
character string, projection
}
\description{
Lambert azimuthal equal area projection string, in old-style PROJ format.
}
\details{
Note that easting/northing are poor names for false X and Y offsets, but
that is what they are called in PROJ.
}
\examples{
laea()
laea(147, -42)
}
