% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{segment}
\alias{segment}
\title{Segment}
\usage{
segment(x = runif(2L, -180, 180), y = runif(2L, -90, 90))
}
\arguments{
\item{x}{longitude (2, start point, end point)}

\item{y}{latitude (2, same as x)}
}
\value{
two-column matrix
}
\description{
A two-point path as a two-column matrix, by default a random one.
}
\examples{
segment()
segment(c(0, 147), c(0, -42))
}
