% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigcurve-package.R
\docType{package}
\name{bigcurve-package}
\alias{bigcurve}
\alias{bigcurve-package}
\title{bigcurve: Adaptive Densifying of Lines}
\description{
Densify lines adaptively, as required by curvature.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/bigcurve}
  \item Report bugs at \url{https://github.com/hypertidy/bigcurve/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Simon Wotherspoon (wrote SGAT::trackMdpts adapted here for midpt) [contributor]
}

}
\keyword{internal}
