% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{segment}
\alias{segment}
\title{Segment}
\usage{
segment(x = runif(2L, -180, 180), y = runif(2L, -90, 90), crs = default_crs())
}
\arguments{
\item{x}{longitude (2, start point, end point)}

\item{y}{latitude (2, same as x)}

\item{crs}{angular projection to use, longlat defaults to 'OGC:CRS84'}
}
\value{
matrix
}
\description{
Segment
}
\examples{
segment()
segment(c(0, 147), c(0, -42))
}
